move rpcHandler into server package where it belongs
