+++ packages/server/src/index.ts
@@ -7,3 +7,4 @@ export {ManifestLoader} from './synchronizer/manifest-loader'
 export {enhance} from './config'
 export {resolveBinAsync} from './resolve-bin-async'
 export {log} from './log'
+export * from './rpc'
